WORKING:  - 수정 중 이미지 저장/삭제 기능 DONE:  - aws 적용 완료  - Review modify, delete  - Review: User's review List  - 리뷰점수들을 합산해서 메뉴에 저장하는 명령 추가 TODO:  - Review: image 동반해서 보내는 기능 추가

+++ LetsEatAll/src/main/java/com/letseatall/letseatall/data/dto/Menu/MenuListDto.java
@@ -10,5 +10,5 @@ public class MenuListDto {
     private String menu_name;
     private int menu_price;
     private String menu_category;
-    private int menu_score;
+    private double menu_score;
 }
